requirements.txt
PyMuPDF==1.26.0
Flask==3.1.1
Werkzeug==3.1.3
tqdm==4.67.1
PyYAML==6.0.2
